+++ api/model.go
@@ -1,0 +1,5 @@
+package api
+
+type Request struct {
+	Url string `json:"url"`
+}
