+++ pkg/storage/database.go
@@ -1,0 +1,3 @@
+package storage
+
+var DB = make(map[string]string)
